Subject domain/db entity implemented

+++ app/src/main/java/com/g3tech/paceful/data/db/daos/StudyDao.kt
@@ -19,7 +19,7 @@ interface StudyDao {
     @Transaction
     suspend fun insertStudyWithTopics(study: Study, topics: List<Topic>) {
         val studyId = insertStudy(study)
-        topics.forEach { topic -> topic.studyId = studyId.toInt() }
+        topics.forEach { topic -> topic.studyId = studyId }
         insertTopics(topics)
     }
