+++ app/src/main/java/com/g3tech/paceful/domain/model/StudyStatus.kt
@@ -1,6 +1,6 @@
 package com.g3tech.paceful.domain.model
 
-enum class StudyStatus(val id: Int) {
+enum class StudyStatus(val id: Long) {
     PENDING(1),
     IN_PROGRESS(2),
     DONE(3),
